Extract RestaurantsContainer
